Conflicts:
	core/java/android/hardware/Camera.java
	core/java/android/os/Process.java
	core/java/android/view/KeyCharacterMap.java
	core/java/android/view/KeyEvent.java
	core/java/android/view/SurfaceView.java
	core/res/res/values/attrs.xml
	core/res/res/values/config.xml
	include/gui/SurfaceTexture.h
	include/gui/SurfaceTextureClient.h
	include/media/stagefright/SurfaceMediaSource.h
	include/ui/KeycodeLabels.h
	libs/binder/Android.mk
	libs/binder/IMemory.cpp
	libs/binder/MemoryHeapBase.cpp
	libs/gui/ISurfaceTexture.cpp
	libs/gui/SurfaceTexture.cpp
	libs/gui/SurfaceTextureClient.cpp
	media/libstagefright/AwesomePlayer.cpp
	media/libstagefright/MediaBuffer.cpp
	media/libstagefright/NuCachedSource2.cpp
	media/libstagefright/StagefrightMetadataRetriever.cpp
	media/libstagefright/SurfaceMediaSource.cpp
	media/libstagefright/codecs/aacenc/basic_op/basic_op.h
	media/libstagefright/matroska/MatroskaExtractor.cpp
	services/java/com/android/server/am/ActivityManagerService.java
	services/surfaceflinger/Android.mk
	services/surfaceflinger/LayerBase.h
	services/surfaceflinger/SurfaceTextureLayer.h

+++ media/libstagefright/include/AwesomePlayer.h
@@ -338,7 +338,7 @@ private:
         int64_t mFirstFrameLatencyStartUs; //first frame latency start
         int64_t mLastFrame;
         int64_t mLastFrameUs;
-        float mFPSSumUs;
+        int64_t mFPSSumUs;
         int64_t mStatisticsFrames;
         bool mVeryFirstFrame;
